Moved methd getCapabilities into Utils
